pyproject.toml
[tool.poetry]
name = "music-sync"
version = "0.1.0"
description = ""
authors = ["Félix Voituret <fvoituret@deezer.com>"]

[tool.poetry.dependencies]
python = "^3.7"
typer = "^0.3.2"
httpx = "^0.18.1"
google-api-python-client = "^2.3.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.4"
google-api-support = "^0.0.4"
pydantic = "^1.8.1"
pytube = "^10.7.2"
rich = "^10.1.0"

[tool.poetry.dev-dependencies]
black = "^21.5b0"
isort = "^5.8.0"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
